otreesurvey/requirements.txt
otree==5.11.1
psycopg2>=2.8.4

# LLM + prompt handling
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=1.10
instructor>=0.3.0
tenacity>=8.2.2
jinja2==2.11.3      
markupsafe==1.1.1